Tests around resetting/failing connections, reuse window tracking
