+++ akka-http2-support/src/test/scala/akka/http/impl/engine/http2/Http2FrameProbe.scala
@@ -63,6 +63,8 @@ private[http2] trait Http2FrameProbe {
   def remainingFromServerWindowFor(streamId: Int): Int
 
   def expectComplete(): Unit
+
+  def within[T](max: FiniteDuration)(f: => T): T = plainDataProbe.within(max)(f)
 }
 
 /**
